Refactor Data
