--- pages/api/studentPortal/quizzes/hireabilityQuiz/hireabilityQuizFE.ts
@@ -1,165 +1,0 @@
-import { QuizViewState } from "../../../../../components/resources/quizzes/shared/types";
-
-export const quizDataFE: QuizViewState = {
-  title: "How hireable are you?",
-  body: "Take this free quiz to reveal your hireability score for software engineering, product, or design roles.",
-  questions: [
-    {
-      title: "FE Question 1",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "FE Question 2",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "FE Question 3",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "FE Question 4",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "FE Question 5",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-  ],
-  currentQuestion: 0,
-  progress: 0,
-};
