--- pages/api/studentPortal/quizzes/hireabilityQuiz/hireabilityQuizME.ts
@@ -1,165 +1,0 @@
-import { QuizViewState } from "../../../../../components/resources/quizzes/shared/types";
-
-export const quizDataME: QuizViewState = {
-  title: "How hireable are you?",
-  body: "Take this free quiz to reveal your hireability score for software engineering, product, or design roles.",
-  questions: [
-    {
-      title: "ME Question 1",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "ME Question 2",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "ME Question 3",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "ME Question 4",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-    {
-      title: "ME Question 5",
-      body: "Select the most relevant options for this question",
-      options: [
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-        {
-          name: "Option",
-          weight: 0,
-          isSelected: false,
-        },
-      ],
-    },
-  ],
-  currentQuestion: 0,
-  progress: 0,
-};
